refactor(api): extract auth routes + fix healthz probe status (#793)
Extract JWT auth endpoints (/api/v1/auth/token, /refresh, /revoke, /me)
into maxwell_daemon/api/routes/auth.py, removing ~150 lines from server.py.
Also fix /healthz to return {"status": "ready"} when backends are available
(the integration test suite expected this and was failing on main).

Closes part of #793
